file placement refactors
